pyproject.toml
[tool.poetry]
name = "ai-backend"
version = "0.0.1"
description = ""
authors = ["Carlos Domani <cad0087@auburn.edu>","Joshua Boyd <jcb0216@auburn.edu>", "Ali Ilkkan Yildirim <aiy0002@auburn.edu>"]

[tool.poetry.dependencies]
python = "^3.9,<3.10"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
gunicorn = "^20.1.0"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg2-binary = "^2.9.1"
alembic = "^1.7.3"
sqlalchemy = "^1.4.25"
arrow = "^1.1.1"
httpx = "^0.19.0"
python-dotenv = "^0.19.0"
aiofiles = "^0.7.0"
Jinja2 = "^3.0.1"
python-multipart = "^0.0.5"
Pillow = "^8.3.2"
sentry-sdk = "^1.4.3"
numpy = "^1.21.4"
opencv-python = "^4.5.4"
tensorflow-cpu = "^2.7.0"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
pytest = "^6.2.5"
pytest-cov = "^2.12.1"
black = "^21.9b0"
isort = "^5.9.3"
sqlalchemy-stubs = "^0.4"

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
